+++ client/app/components/queries/QueryEditor/index.jsx
@@ -1,47 +1,18 @@
 import React from "react";
 import PropTypes from "prop-types";
-import Tooltip from "antd/lib/tooltip";
 import { react2angular } from "react2angular";
-
-import AceEditor from "react-ace";
-import ace from "brace";
-import notification from "@/services/notification";
-
-import "brace/ext/language_tools";
-import "brace/mode/json";
-import "brace/mode/python";
-import "brace/mode/sql";
-import "brace/mode/yaml";
-import "brace/theme/textmate";
-import "brace/ext/searchbox";
-
+import Tooltip from "antd/lib/tooltip";
+import AutocompleteToggle from "@/components/AutocompleteToggle";
 import { DataSource, Schema } from "@/components/proptypes";
 import { Query } from "@/services/query";
 import { QuerySnippet } from "@/services/query-snippet";
 import { KeyboardShortcuts } from "@/services/keyboard-shortcuts";
-
+import notification from "@/services/notification";
 import localOptions from "@/lib/localOptions";
-import AutocompleteToggle from "@/components/AutocompleteToggle";
 
+import { AceEditor, langTools, snippetsModule } from './ace';
 import { buildKeywordsFromSchema } from "./utils";
 import "./index.less";
-
-const langTools = ace.acequire("ace/ext/language_tools");
-const snippetsModule = ace.acequire("ace/snippets");
-
-// By default Ace will try to load snippet files for the different modes and fail.
-// We don't need them, so we use these placeholders until we define our own.
-function defineDummySnippets(mode) {
-  ace.define(`ace/snippets/${mode}`, ["require", "exports", "module"], (require, exports) => {
-    exports.snippetText = "";
-    exports.scope = mode;
-  });
-}
-
-defineDummySnippets("python");
-defineDummySnippets("sql");
-defineDummySnippets("json");
-defineDummySnippets("yaml");
 
 class QueryEditor extends React.Component {
   static propTypes = {
